Cleanup
Huge cleanup. Removal of several archaic smart pointers and modernization of others. Turned on compiling with C++17. Some work to remove usage of the most deprecated system APIs. Fixed a bug with autoclosing output window deleting the handler object twice. Removal of 32-bit and Carbon code. Removal of ICU Regex which is not public and works on 2-byte Unicode in favor of UTF-8 regex. Added build.sh as a starting point for building all components.