Remove OMCDeputy and fix Suport symlink
Support symlink in Abracode.framework was broken in previous cleanup change where the location of the build output changed
OMCDeputy was added many years ago for situation  when a commnad was executed synchronously within a host and you wanted to send an Apple Event (via Apple Script) to the host - this resulted in a deadlock. The problem was mainly in Finder when contextual menus were loaded in-proc. Since Apple no longer loads CMs in 64-bit apps this is not an interesting scanario. Furthermore, if you really need to send AppleEvent to the host which executes your command you can choose non-synchronous execution like popen and use osascript to execute your AppleScript. The workaround is easy enough that we don't need complex support for this in OMC engine itself - let's remove it now.