Always export environment variables for all nib dialog controls
All environment variables with nib control values are now exported for popen execution.
The only exception is OMC_NIB_TABLE_N_COLUMN_M_ALL_ROWS which is rarely used and expensive to obtain. It must be explicitly requested by putting it in the body of the command in plist or in ENVIRONMENT_VARIABLES dictionary
While at it, added sanitization of the control id/tag. Negative tags are not valid. Only alphanumeric characters in control IDs are valid. This filters out not interesting controls which don't have IDs assigned therefore the script will never use them