Add OMC_APP_BUNDLE_PATH and deprecate OMC_MY_HOST_BUNDLE_PATH
OMC_MY_HOST_BUNDLE_PATH made most sense in the plug-in world when plug-ins were loaded in-proc. This carried-over to applets (aka droplets) but really the "host" concept does not apply so a better name for it is OMC_APP_BUNDLE_PATH
So starting with OMC 4.0 we deperecate OMC_MY_HOST_BUNDLE_PATH (still availble and working as before) but prefer OMC_APP_BUNDLE_PATH, which is also always exported as env variable for popen executions unlike the old env variable. That way we don't break backwards compatibility