Add entitlements to apps
Apps like CommandDroplet and OMCService need entitlements to load the Abracode.framework. With hardened runtime this will be needed if you re-codesign the app but still need to load Abracode.framework which will be signed for distribution by different entity.
Also, specify that we want to send AppleEvents to other apps.
I am not sure about other entitlements. With CommandDropet being a generic template for the apps to be built with scripts by users we may want to allow more (?). This is baked into exeuctable binary at build time. I still need to see what will be the experience of building an applet for running in Catalina - hopefully it will not require building an applet in Xcode - just re-codesigning.