Add support for finding Abracode.framework in Frameworks folder next to the applet
Command applets (droplets) can now find Abracode.framework in external "Frameworks" folder next to applet(s) in addition to "Frameworks" in app bundle and "Frameworks" in /Library. This makes it convenient to deploy updates to Abracode.framework without forcing users to go through the installer.
When building the AbracodeFramework project: create "Frameworks" folder in build location and put the symlink to Abracode.framework there so CommandDroplet and test app CocoaExecutor can find it.