Remove legacy encodeWithCoder: in all OMC controls
"initWithCoder:" and "encodeWithCoder:" was used long time ago by IB plugins to serialize custom control properties to nibs. With that mechanism long gone, remove that code. This will break if there is any applet with old nib file with embedded properties - this would need to be opened in current Xcode and updated or converted with ibtool.