Bring back initWithCoder: in minimal form and remove encodeWithCoder: for OMC controls
Removal of encodeWithCoder: and initWithCoder: introduced a bug that some properties were not set to their intended default values. Objective-C ivars are zeroed out in alloc and any values intended to be non-0, no-nil, non-NO at init must be explicitly set. I thought initWithCoder: is not used but actually it is, even if no custom values are persisted in nibs. Therefore bring back initWithCoder: intializers with intended defaults. At the same time I decided to go further and simplify properites in this code. I only stopped short of changing it to all ARC. This requires a separate pass and special attention b/c it is very easy to get it wrong and end up with crashes.