Add omc_table_prepare_empty special instruction for omc_dialog_control
omc_table_prepare_empty is an instruction to send to OMCTableView to clear the table by removing all data but not reload the table yet. This eliminates the flicker for tables when a quick succession of fast removal and repopulation is actually visible. When the retrieval of the new data takes a while, refresh after cleaning the table is desired to indicate in UI that the new content is coming and not leave old stale data - in these cases you should still use omc_table_remove_all_rows.
I did some additional cleanup in omc_dialog_control code. For fun I removed almost all memory deallocations and CFReleases. Now all these objects are leaking! So sue me :-) If you can explain why it is OK, you can probably pass my interview for a software engineer position.