Implement new execution type: exe_script_file and exe_script_file_with_output_window
The new execution style is to simplify development of applets. The idea is to allow each command/handler be in a separate script in Applet.app/Contents/Resources/Scripts and not in Command.plist which may grow unpleasantly big and you have to deal with XML escaping. The connection between command and the script is the command ID. If you name the script the same as your COMMAND_ID and drop it into Resources/Scripts folder it will be found automatically for you. The extension you give to the scriot will determine the shell which will execut it. If you want to customize the shell, yo ustill can with POPEN_SHELL setting. Combined with automatic export of most useful OMC environment variables this allows to have a cleaner, more readable and maintainable applet development setup.
When using exe_script_file or exe_script_file_with_output_window it is rcommended you delete the <key>COMMAND</key> completely. On rare occasions you want to control where the script is located you can fill the COMMAND with absolute path like this:
		<key>COMMAND</key>
			<array>
				<string>__APP_BUNDLE_PATH__</string>	<string>/Contents/Resources/MyLocation/MyScript.sh</string>
			</array>
Special OMC words in form of  __SPECIAL_WORD__ are accepted in the path. I expect that custom script location is usually not going be wildely used but the implementation was easy enough to give that flexibility if needed.