Add support for App Group Identifier for CFMessagePort communication
When creating CFMessagePort for communication between omc_dialog_control and the host (which code is in Abracode.framework) we need to prefix the port name with app group identifier, which is apple developer team identifer. It is set in OmcDistribution.xcconfig as DEVELOPMENT_TEAM. Open source version has this value empty but for binary distribution it is the team identifier issued by Apple with codesigning certificate
Added a script phase to omc_dialog_control to generate AppGroupIdentifier.h header in build directory, which is included by both omc_dialog_control and Abracode framework projects.