Add supportfor debugging if control keyboard modifier is held
Sometimes it is hard to figure our what is going wrong with a command script and logging in output window is helpful to debug it. Adding a "control" keyboard modifier to make it easier - no need to change the comamnd execution mode in the plist - just hold the control key when executing the command and the silent modes will change into "_with_output_window" execution modes.
We might be able to do more with the debugging mode in the future, like logging the internal flow and errors in some console - for now this is the bare bones implementation.