Change default location for contextual menu items to be root
When OMC was developed first there was a vibrant ecosystem of contextual menu plug-ins for Mac and many people had multiple plug-ins installed so there was a need to organize them. By default "On My Command" was the submenu for new commands/items. It is not needed now. It is more convenient to have the menu items be right in the root menu. You can still put them in submenu if you desire so by using <key>SUBMENU_NAME</key>.

The only way to load 64-bit contextual menu plug-ins now is via Shortcuts 2.1:
http://www.abracode.com/free/cmworkshop/shortcuts.html
Shortcuts has been updated and is now codesigned and notarized for macOS 10.15 Catalina