Tying loose ends for distribution of OMC 4.0
OMC workspace is needed for "Archive" operation because of Xcode bug: when archiving Xcode is not finding binaries produced by different projects to copy inside final bundle of other projects.
Added "OMC Distribution" scheme to produce Release version of all distributed articacts, ready for notarization.
Added export file to OnMyCommandCM plug-in with the entry symbol and hide everything else.
All command line tools are now built with "dwarf-with-dsym" debug information format and .dSYM files are produced in Release build. Other binaries, most notably Abracode.framework, are produced with embedded symbols so if there is a crash, we will see the symbols in the report.