Add processing of pending events and notificaitons before executing next command
A deficiency has been noted that next command was not getting the control values set in previous command. It is because the runloop never got a chance to process them - it is a single thread design which relies on CFRunLoop.
The fix is to run the runloop until all sources are processed before starting the next command.