Add OMCGridView
OMCGridView is a replacement for NSGridView which adds a "tag" custiom property and allows enabling and disabling child views. That way a tag can be set for this view in interface builder a sa custom integer property and a single omc_enable/omc_disable messaeg can disable all subviews. The implementation is cloned from OMCBox