Create new OMC control: OMCWebKitView
WebKit's WebView control is deprecated. OMC needs a new control based on WKWebView. OMCWebKitView replaces deprectated OMCWebView.
The design is that OMCWebKitView inherits from NSView and not WKWebView. When WKWebView is instantiated from nib it cannot be configured. For some strange reason Apple chose the design that the configuration cannot be changed once the control is created. It effectively forces everyone to create WKWebView in code so this is what OMCWebKitView does in awakeFromNib.
This is the first version just to set up the basics.