Add JavaScript injection to OMCWebKitView
When OMCWebKitView is instantiated, the OMC engine first injects the OMCWebKitSupport.js script with instruction to load it at document start, before DOM is finished loading. This script is meant to be a library and contain support scripts for OMC functionality and most likely no code will actually be executed at this point. Second, if a client/dialog nib control specified "javaScriptFile" key+value in "User Defined Runtime Attributes" in Interface Builder, that string is meant to be the name of the Java Script to be injected in the WKWebView as well. This name specified in  "javaScriptFile" should be without .js suffix but if client does specify name with .js, it will be handled properly. No other suffixes are supported. If you specify javaScriptFile=ExampleScript, the script file is expected to be Contents/Resources/Scripts/ExampleScript.js in the applet. See example usage in OMCTestApp. That script is injected at the document end, after loading and allows the client to post-process the DOM or do whatever it needs to do. That script will have OMC library APIs available for elements manipulation. Example usage is that when loading HTML content which you don't have control over you can post-process elements and turn them into OMC elements and the engine will automatically retrieve values from such elements or trigger commands with chosen IDs when control values change. That functionality is planned but not implemented yet.