Fix regression when getting control values from modal dialogs
The dialog proxy object OMCDialog was being destroyed after modal dialog dismissal. This proxy object holds the control values obtained from the dialog and was inaccessible after the window was closed. This change makes OMCDialog refcounted and prolongs the lifetime of the proxy object to make the control values available when needed.