Table view selection fixes
Row selection was incorrect (cached from last selection) if the current selection was empty. The control value store should remove the previous value if we deselect all control parts.

Double-click on the table view in the empty area without valid row selection was triggering the double-click command - which is problematic if a command has an output window or a dialog associated with it and no selection data to use. The intent is to have a double-click command only for a valid row selection.