Add support for "main" shell script
When a command does not have an explicitly defined COMMAND_ID it is assumed to be a "top" or main command, while others with the same NAME are considered sub-commands.
For exe_script_file & exe_script_file_with_output_window this change allows finding main.sh in "$OMC_APP_BUNDLE_PATH/Contents/Resources/Scripts" location, which would be considered the main applet entry. Of course, it will not scale if there are multiple unique commands in the Command.plist and they all would use the same main.sh script. But the idea is that there can only be one main entry point associated with the key droplet command which is executed automatically on app launch.