Switch from Obj-C manual refcounting to ARC
The code started its life before ARC was invented and somehow had never been modernized. This is a long overdue refactoring. Because of heavy usage of CoreFoundation objects, a lot of bridging casts had to be added. If something is not done correctly it potentially may result in leaks or crashes. Passed basic sanity tests and the testing will continue.