Separate command description, input context and command runtime data
The goal of this change is to enable associating one context file with one dialog. Previous behavior was that the dialog was always shown before the decision was made to process file objects together or separately - controlled by PROCESSING_MODE: proc_separately|proc_together (separate processing is default). The new desired behavior is that processing a command with a dialog and multiple file objects will result in multiple dialogs - many as there are input files. If you wish to have a command which presents a single dialog for multiple objects together you would need set PROCESSING_MODE: proc_together in command plist.
That seemingly simple goal turned into a big change and refactoring due to limitations of original contextual menu plug-in design. Let's take this opportunity to describe that design.
OnMyCommandCM class is the core of the contextual menu plug-in and its functionality is to examine context, dynamically add menu items for matching context and execute selected command. OnMyCommandCM used to hold all functionality and data in one class and one object at runtime. It is now broken out into:
- CommandDescription - this contains all settings read from Command.plist. OnMyCommandCM holds a list of all these and still is the central point of dispatching the commands.
- OMCContextData - this object holds the input file objects or selected text. In an applet, this could be files dropped on the app or selected with a navigation dialog. In the system service it is selection similar to original contextual menu plugins.
- CommandRuntimeData - this object is associated with a running command and contains OMCContextData. OMCContextData is passed from one command's CommandRuntimeData to subcommand's on next command's CommandRuntimeData. CommandRuntimeData is not tied to OnMyCommandCM object anymore so there may exists more of these at runtime per one single logical command.
This separation of objects allows splitting one logical command with multiple objects into runtime instances of the same command but different private CommandRuntimeData and private OMCContextData.