Generate some up-to-date documentation with Grok
Starting with reference files for command description:
omc_command_reference.md
omc_runtime_context_reference.md
The work is not complete yet but a majority of most important options should be described.
Also, Grok heavily hallucinates, imagining options which are not there and needs to be constantly supervised and corrected. Hope none of these slipped through in review.