Small tweak for finding script files for exe_shell_script
Command group's main command handler description may not have a `COMMAND_ID` declared. In the absence of command id, the OMC engine was looking for "main.sh" in the Content/Resources/Scripts/.
This change makes it look first for `<CommandName>.main.sh` first and then fall back to main.sh if not found, where `<CommandName>` is the non-dynamic `NAME` value in command description. If your command constructs dynamic menu names, such name is ignored for the purpose of searching for script in applet bundle.