Propagate parent context to subcommand context check
Initially CommonContextCheck() was not considering the parent command context - only ExecuteCommand() was inheriting that. But if you add some context check in plist description of a subcommand then CommonContextCheck() fails and prevents execution. For consistency and to avoid nasty surprises why the subcommand fails to execute let's propagate the parent context info to CommonContextCheck() as well.